Lots of upgrades.
Camera controls are less invasive at the data layer now control is handled in the ViewModel. Camera controls are now an attached behavior which promotes sharing of the mouse logic with any control. D3D view is now implemented with complete rendering using vertexbuffers and textures cached at the MementoServer.