Switch to iterative BVH walking.
We connect all nodes in the BVH with a child pointer and a "next node" pointer. The next node is either the next child or the sibling of the parent. In this manner we can walk child first and next last which forms a complete walk through all nodes. This makes walking the BVH iterative which dramatically increases performance.