More taps for the desktop shader.
"Unlit" is a bad name since this shader is being used for the desktop projection in VR. Taking one sample of a single MIP 0 texture is going to produce a result just as bad as you'd expect, along with significant (and highly distracting) shimmering. Instead use ddx/ddy to compute a trivial box filter.