Add in textures for Vulkan.
Unlike D3D12, Vulkan is annoying as all hell and requires a bonkers amount of code to perform even the simplest operations. Oh well. We try to exploit the flexibility of Vulkan to mimic the operation of D3D12 which includes NOT using combined sampler/images. Unfortunately this makes the ugly binding code look even uglier.