Farewell raytracer...
The raytracer in here was fun to get going, and we managed to get it going in both CUDA and AMP. Unfortunately for us CUDA doesn't work in UWP and AMP is horrible. The remaining software raytracer is too slow to be useful. If we want raytracing back then the way to do it is DXR.